fix: 코드리뷰 반영 및 전체width 확장
